Got HHN working
It only works with smaller networks (<5000 edges at least and >4 nodes). We may need to introduce a SPRASlite or different version for this to work. As of right now HHN does not seem like it should be implemented into SPRAS given how long it takes to work with networks of reasonable size (more than 5000 edges).